+++ dev/sg/internal/run/run.go
@@ -3,7 +3,6 @@ package run
 import (
 	"context"
 	"crypto/md5"
-	"encoding/json"
 	"fmt"
 	"io"
 	"os"
@@ -12,13 +11,11 @@ import (
 	"runtime"
 	"strings"
 	"sync"
-	"syscall"
 	"time"
 
 	"github.com/grafana/regexp"
 	"github.com/rjeczalik/notify"
 	"golang.org/x/sync/semaphore"
-	"k8s.io/utils/strings/slices"
 
 	"github.com/sourcegraph/sourcegraph/dev/sg/internal/analytics"
 	"github.com/sourcegraph/sourcegraph/dev/sg/internal/std"
@@ -30,111 +27,6 @@ import (
 )
 
 const MAX_CONCURRENT_BUILD_PROCS = 4
-
-type pidFile struct {
-	Args []string `json:"args"`
-	Pid  int      `json:"pid"`
-}
-
-func PidExistsWithArgs(args []string) (int, bool, error) {
-	homeDir, err := os.UserHomeDir()
-	if err != nil {
-		return 0, false, errors.Wrap(err, "could not check pidfiles")
-	}
-
-	pattern := filepath.Join(homeDir, ".sourcegraph", "sg.pid.*")
-	matches, err := filepath.Glob(pattern)
-	if err != nil {
-		return 0, false, errors.Wrap(err, "could not list pidfiles")
-	}
-
-	for _, match := range matches {
-		f, err := os.Open(match)
-		if err != nil {
-			return 0, false, errors.Wrapf(err, "could not check pidfile %q", match)
-		}
-		defer f.Close()
-
-		var content pidFile
-		if err := json.NewDecoder(f).Decode(&content); err != nil {
-			return 0, false, errors.Wrapf(err, "could not check pidfile %q", match)
-		}
-
-		if slices.Equal(content.Args, args) {
-			// If a pid already exists, let's check if it's still running.
-			alive, err := isPidAlive(int32(content.Pid))
-			if err != nil {
-				return 0, true, errors.Wrapf(err, "could not check if pid %d is alive", content.Pid)
-			}
-			if !alive {
-				// Trash the pidfile and return that it's not running.
-				_ = os.Remove(match)
-			}
-			return content.Pid, true, nil
-		}
-	}
-
-	return 0, false, nil
-}
-
-// On Unix systems, os.FindProcess always returns a os.Proc, regardless if the process is running or not. Therefore,
-// we need more work to check if it's alive or not.
-// Reference: https://github.com/shirou/gopsutil/blob/c141152a7b8f59b63e060fa8450f5cd5e7196dfb/process/process_posix.go#L73
-func isPidAlive(pid int32) (bool, error) {
-	if pid <= 0 {
-		return false, fmt.Errorf("invalid pid %v", pid)
-	}
-	proc, err := os.FindProcess(int(pid))
-	if err != nil {
-		return false, err
-	}
-	err = proc.Signal(syscall.Signal(0))
-	if err == nil {
-		return true, nil
-	}
-	if err.Error() == "os: process already finished" {
-		return false, nil
-	}
-	errno, ok := err.(syscall.Errno)
-	if !ok {
-		return false, err
-	}
-	switch errno {
-	case syscall.ESRCH:
-		return false, nil
-	case syscall.EPERM:
-		return true, nil
-	}
-	return false, err
-}
-
-func writePid() error {
-	homeDir, err := os.UserHomeDir()
-	if err != nil {
-		return err
-	}
-
-	pidFileName := fmt.Sprintf("sg.pid.%d.json", os.Getpid())
-	pidFilePath := filepath.Join(homeDir, ".sourcegraph", pidFileName)
-
-	content := pidFile{
-		Args: os.Args[1:],
-		Pid:  os.Getpid(),
-	}
-
-	b, err := json.Marshal(content)
-	if err != nil {
-		return err
-	}
-
-	if err := os.WriteFile(pidFilePath, b, 0644); err != nil {
-		return err
-	}
-
-	interrupt.Register(func() { os.Remove(pidFilePath) })
-
-	return nil
-}
 
 func Commands(ctx context.Context, parentEnv map[string]string, verbose bool, cmds ...Command) error {
 	chs := make([]<-chan struct{}, 0, len(cmds))
